Component date range
